src/reports/requirements.txt
asn1crypto==0.24.0
attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cryptography==2.6.1
entrypoints==0.3
idna==3.3
isodate==0.6.1
keyring==17.1.1
keyrings.alt==3.1.1
lxml==4.9.1
platformdirs==2.5.2
pycrypto==2.6.1
PyGObject==3.30.4
python-dotenv==0.20.0
pytz==2022.1
pyxdg==0.25
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
SecretStorage==2.3.1
six==1.12.0
tabulate==0.8.10
urllib3==1.26.10
xmltodict==0.13.0
zeep==4.1.0
